+++ src/main/java/com/innowise/demo/serserviceapp/service/UserService.java
@@ -4,12 +4,13 @@ import com.innowise.demo.serserviceapp.dto.UserDto;
 import com.innowise.demo.serserviceapp.mapper.UserMapper;
 import com.innowise.demo.serserviceapp.model.User;
 import com.innowise.demo.serserviceapp.repository.UserRepository;
-import jakarta.validation.Valid;
+import com.innowise.demo.serserviceapp.exception.UserNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 
 import java.util.Optional;
 
@@ -28,31 +29,34 @@ public class UserService {
     }
 
     public Optional<UserDto> getUserById(Long id) {
-        return userRepository.findById(id).map(userMapper::toDto);
+        User user = userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User not found with id: " + id));
+        return Optional.of(userMapper.toDto(user));
     }
+
     public Page<UserDto> getAllUsers(Pageable pageable) {
         return userRepository.findAll(pageable).map(userMapper::toDto);
     }
+
     public Optional<UserDto> getUserByEmail(String email) {
         return userRepository.findByEmail(email).map(userMapper::toDto);
     }
 
-    public UserDto updateUser(Long id, @Valid UserDto userDetails) {
-        Optional<User> optionalUser = userRepository.findById(id);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
-            user.setName(userDetails.getName());
-            user.setSurname(userDetails.getSurname());
-            user.setBirthDate(userDetails.getBirthDate());
-            user.setEmail(userDetails.getEmail());
-            User savedUser = userRepository.save(user);
-            return userMapper.toDto(savedUser);
-        }
-        return null;
+    @Transactional
+    public UserDto updateUser(Long id, UserDto userDetails) {
+        User user = userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User not found with id: " + id));
+        user.setName(userDetails.getName());
+        user.setSurname(userDetails.getSurname());
+        user.setBirthDate(userDetails.getBirthDate());
+        user.setEmail(userDetails.getEmail());
+        User savedUser = userRepository.save(user);
+        return userMapper.toDto(savedUser);
     }
 
     @Transactional
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
